refactor: move db creation to factory fn for env-independent instantiation
